New version of madx and some minor modifications
1. Updated the madx version to the most recent ones. This avoided a bug
that I encountered earlier, where a negative drift was reported by madx
and the program terminated with an error. Only the linux version of the
binary is updated, since I cannot test the mac version.
2. For better debugging the error messages of madx (stderr) are now
passed as part of the output. Before, only the stdout was part of the
output.